First commit. This implements debevec along with reinhard02.
There might be errors! Need to read the reinhard paper again and confirm I got it right.

Inorder to run, Use the following command in matlab -
HDR('../Images','.jpg') where Images is the Images folder and jpg is the extension of the images in the folder. Make sure you only have image files in the folder. Nothing else.